fix(ui): move remaining marketplace tab features to its own section
Some variables, mocked data from the previous marketplace tab were overlooked. They have been placed
in the correct folders.

fix #452
