+++ app/renderer/ui/components/main/sections/wallet/MockData.ts
@@ -5,49 +5,6 @@ export const balanceData = {
   total: "4.241592"
 }
 
-export const marketplaceProducts = [
-  {
-    title: "n-of-m multisig",
-    author: "aesedepece",
-    cover: "https://os.alipayobjects.com/rmsportal/QBnOOoLaAfKPirc.png",
-    alt: "example",
-    rating: "4.5",
-    tags: ["multisig", "parametric"]
-  },
-  {
-    title: "Pay to Alice or Bob, based on API response",
-    author: "aesedepece",
-    cover: "https://os.alipayobjects.com/rmsportal/QBnOOoLaAfKPirc.png",
-    alt: "example",
-    rating: "4.5",
-    tags: ["bet", "oracle", "parametric"]
-  },
-  {
-    title: "Crypto-heirship",
-    author: "aesedepece",
-    cover: "https://os.alipayobjects.com/rmsportal/QBnOOoLaAfKPirc.png",
-    alt: "example",
-    rating: "4.5",
-    tags: ["oracle", "parametric"]
-  },
-  {
-    title: "Pay if flight is delayed",
-    author: "aesedepece",
-    cover: "https://os.alipayobjects.com/rmsportal/QBnOOoLaAfKPirc.png",
-    alt: "example",
-    rating: "4.5",
-    tags: ["oracle", "parametric"]
-  },
-  {
-    title: "Pay if it rains",
-    author: "aesedepece",
-    cover: "https://os.alipayobjects.com/rmsportal/QBnOOoLaAfKPirc.png",
-    alt: "example",
-    rating: "4.5",
-    tags: ["oracle", "parametric", "weather"]
-  }
-]
-
 export const selectFeesData = [
   {
     fee: 79,
